requirements.txt
flask==3.0.3
openai==1.30.1
httpx>=0.23.0,<0.28.0
python-dotenv==1.0.1
